now exporting two classes, DatePicker and DateRangePicker
split the logic responsible for the calendar and for the selection in two pieces. Therefor now it is possible to come up with other selection logic on top of the calendar implementation.

refactored the MonthProvider, because the render function should be without side effects

introduced exports.js as entry point for the bundler

reworked the componentDidUpdate logic of daterangepicker, not quite sure, whether the behaviour is as expected, might need to change some things there.

changed "PropTypes" to "propTypes" (consistency)

closes #6