added id to props of date(range)picker
if no id is passed in, the behaviour will be the same like before.
If another id, compared to the previous props, is passed in, you can force a focus even with the same selection.

fixed testcase, which failed due to the precision tag in momentjs function call (momentjs will floor values like 0.999 to 0, as far as I could determine)